requirements.txt
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
responses==0.23.3
